Refactor pin mobile view
